requirements.txt
python-can>=4.3.1
pyserial>=3.5
PyQt6>=6.6.0